Docker/Dockerfile-Python/requirements.txt
Flask == 2.2.2
Flask-MySQLdb == 1.0.1
mysqlclient==2.1.1
SQLAlchemy == 1.4.42
psycopg2_binary == 2.9.4
Flask_Migrate == 3.1.0
pymongo == 4.0.1